+++ packages/riverpod_generator/test/integration/sync.dart
@@ -88,6 +88,10 @@ String _private(_PrivateRef ref) {
 /// This is some documentation
 @riverpod
 class PublicClass extends _$PublicClass {
+  PublicClass([this.param]);
+
+  final Object? param;
+
   @override
   String build() {
     return 'Hello world';
@@ -107,6 +111,10 @@ class _PrivateClass extends _$PrivateClass {
 /// This is some documentation
 @riverpod
 class FamilyClass extends _$FamilyClass {
+  FamilyClass([this.param]);
+
+  final Object? param;
+
   @override
   String build(
     int first, {